To avoid polluting the Schema.org implementation with our assumptions, schemas that are represented by non-post WordPress objects will instead use WP-namespaced objects.
For example, the Schemify\Schemas\WP\User class is a representation of a WP_User object, but conforms to the Person schema. This enables us to say, for instance, that a WP\User is the author of a CreativeWork *without* forever restricting the Person schema to be a WordPress user (for example, if there's a post type of people with profiles, for which 'Person' would be the appropriate schema).

All WP schemas should use the WP_Schema trait, which will prevent this work-around from introducing invalid Schemas.